fix install linting script

+++ install_linting.sh
@@ -6,7 +6,9 @@ SLN_ROOT=$1
 
 cp Directory.Build.props.template $SLN_ROOT/Directory.Build.props
 cat .editorconfig.template >> $SLN_ROOT/.editorconfig
-cp lint.sh $SLN_ROOT/
+if [ ! -f "$SLN_ROOT/lint.sh" ]; then
+    cp lint.sh $SLN_ROOT/
+fi
 pushd $SLN_ROOT
 dotnet new tool-manifest
 dotnet tool install csharpier
